mod: AddChild page and ViewModel move to its folder
