+++ VaxineApp/VaxineApp/Views/Home/Family/Child/AddChildPage.xaml
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="VaxineApp.Views.Home.Family.AddChildPage"
+             x:Class="VaxineApp.Views.Home.Family.Child.AddChildPage"
              xmlns:local="clr-namespace:VaxineApp.Renderers"
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:pages="clr-namespace:VaxineApp.Views.Home.Family"
+             xmlns:pages="clr-namespace:VaxineApp.Views.Home.Family.Child"
              Shell.TabBarIsVisible="False"
              Shell.PresentationMode="Modal"
              Title="Add Child">
